// vf-os-vapp-3.4-master/vapp4/src/app/api/model/createControlInformation.ts
/**
 * vf-OS vApp 4.3 Endpoints
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0.0
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface CreateControlInformation {
  cps?: any;
  mo?: any;
  ope?: any;
  product?: any;
  product_cost?: any;
  ressource_cost?: any;
  control_type?: any;
  control_size?: any;
  measure?: any;
  max_tolerence?: any;
  min_tolerence?: any;
  customer?: any;
}