// vf-os-vapp-3.4-master/vapp4/src/app/api/model/manufacturingOrder.ts
/**
 * vf-OS vApp 4.3 Endpoints
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0.0
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface ManufacturingOrder {
  numofs?: number;
  codcpt?: string;
  libar1?: string;
  qtepre?: number;
}